+++ example/lib/model/frame_model.dart
@@ -21,7 +21,7 @@ class FrameModel extends AbsExModel {
 
   @override
   List<Object?> get props => [...super.props, name, width, height, posX, posY, angle, bgColor];
-  FrameModel() : super(type: ExModelType.frame, parent: '') {
+  FrameModel(String pmid) : super(pmid: pmid, type: ExModelType.frame, parent: '') {
     name = UndoAble<String>('', mid);
     bgUrl = UndoAble<String>('', mid);
     width = UndoAble<double>(0, mid);