+++ lib/hycop/enum/model_enums.dart
@@ -1,6 +1,5 @@
 //import 'dart:ui';
 
-
 /////////////////
 // <!-- 삭제될 예정
 enum ExModelType {
@@ -9,13 +8,14 @@ enum ExModelType {
   page,
   frame,
   contents,
+  user,
+  team,
   end;
 
   static int validCheck(int val) => (val > end.index || val < none.index) ? none.index : val;
   static ExModelType fromInt(int? val) => ExModelType.values[validCheck(val ?? none.index)];
 }
 ///////////////////-->
-
 
 enum AccountSignUpType {
   none,
@@ -53,7 +53,7 @@ enum ContentsType {
   static int validCheck(int val) => (val > end.index || val < none.index) ? none.index : val;
   static ContentsType fromInt(int? val) => ContentsType.values[validCheck(val ?? none.index)];
   static getContentTypes(String contentType) {
-    if(contentType.contains("image")) {
+    if (contentType.contains("image")) {
       return ContentsType.image;
     } else if (contentType.contains("video")) {
       return ContentsType.video;